feat:Permission has been addeed to actions without permission

+++ src/controllers/web/ItemController.php
@@ -189,7 +189,9 @@ class ItemController extends Controller
 
     public function actionClone()
     {
-
+        if (!\Yii::$app->user->can('menuWebItemClone')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         if (Yii::$app->request->isAjax) {
             $id_menu = Yii::$app->request->post('DynamicModel')['id_menu'];
             $id_item = Yii::$app->request->post('id_item');
@@ -208,47 +210,51 @@ class ItemController extends Controller
 
     public function actionMove()
     {
+        if (!\Yii::$app->user->can('menuWebItemMove')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
+
         if (Yii::$app->request->isAjax) {
             $id_menu = Yii::$app->request->post('DynamicModel')['id_menu'];
             $id_item = Yii::$app->request->post('id_item');
-    
+
             try {
                 $id_parent = Yii::$app->request->post('DynamicModel')['id_parent'];
             } catch (\Throwable $th) {
                 $id_parent = null;
             }
-    
-            
+
+
             $menuModel = Menu::findOne($id_menu);
             if (!$menuModel) {
                 return $this->asJson(['status' => 'error', 'message' => 'Menu not found.']);
             }
-    
-            
+
+
             $existingRelation = ItemChild::find()->where(['id_child' => $id_item])->one();
-    
-            
+
+
             if ($existingRelation) {
                 $existingRelation->delete();
             }
-    
-            
+
+
             if ($id_parent !== null && $id_parent != 0) {
                 $newRelation = new ItemChild();
                 $newRelation->id_item = $id_parent;
                 $newRelation->id_child = $id_item;
-    
+
                 if (!$newRelation->save()) {
                     return $this->asJson(['status' => 'error', 'message' => 'Failed to move item.']);
                 }
             }
-    
+
             return $this->asJson(['status' => 'success']);
         }
-    
+
         return $this->asJson(['status' => 'error', 'message' => 'Invalid request.']);
     }
-    
+
 
 
     private function deleteItem($id_item, $id_parent = null, $id_menu = null)
@@ -396,52 +402,52 @@ class ItemController extends Controller
     public function actionParentList()
     {
         $out = [];
-    
+
         if ($this->request->isPost) {
             $request = $this->request->post('depdrop_parents');
             $id_menu = isset($request[0]) ? $request[0] : null;
-    
+
             if ($id_menu == null || $id_menu == '') {
                 return json_encode(['output' => [], 'selected' => '']);
             }
-    
+
             $menu = Menu::findOne($id_menu);
             $menuItems = $menu->getItems()->orderBy(['sort' => SORT_ASC, 'id_item' => SORT_ASC])->asArray()->all();
-    
-            
+
+
             $menuRelations = ItemChild::find()->asArray()->all();
-    
-            
+
+
             $menuTree = [];
             foreach ($menuRelations as $relation) {
                 $menuTree[$relation['id_item']][] = $relation['id_child'];
             }
-    
-            
+
+
             $sortedItems = [];
             foreach ($menuItems as $item) {
                 $sortedItems[$item['id_item']] = $item;
             }
-    
+
             $rootParents = array_diff(array_keys($sortedItems), array_column($menuRelations, 'id_child'));
-    
+
             $out[] = ['id' => 0, 'name' => Module::t('Root')];
-    
+
             function buildParentChildTree($parentId, $sortedItems, $menuTree, $depth = 0, $forMove = false)
             {
                 $result = [];
                 if (isset($menuTree[$parentId])) {
-    
+
                     usort($menuTree[$parentId], function ($a, $b) use ($sortedItems) {
                         return ($sortedItems[$a]['sort'] ?? 0) - ($sortedItems[$b]['sort'] ?? 0);
                     });
-    
+
                     foreach ($menuTree[$parentId] as $childId) {
                         if (isset($sortedItems[$childId])) {
                             $childItem = $sortedItems[$childId];
-    
+
                             $indentation = ($depth > 0 && $forMove) ? str_repeat('- ', $depth) : '';
-    
+
                             $result[] = ['id' => $childItem['id_item'], 'name' => $indentation . $childItem['label']];
                             $result = array_merge($result, buildParentChildTree($childItem['id_item'], $sortedItems, $menuTree, $depth + 1, $forMove));
                         }
@@ -449,35 +455,35 @@ class ItemController extends Controller
                 }
                 return $result;
             }
-    
+
             $menuHierarchicalList = [];
             foreach ($rootParents as $parentId) {
                 if (isset($sortedItems[$parentId])) {
                     $parentItem = $sortedItems[$parentId];
-    
+
                     $menuHierarchicalList[] = ['id' => $parentItem['id_item'], 'name' => $parentItem['label']];
                     $menuHierarchicalList = array_merge($menuHierarchicalList, buildParentChildTree($parentItem['id_item'], $sortedItems, $menuTree, 1, false));
                 }
             }
-    
+
             $moveHierarchicalList = [];
             foreach ($rootParents as $parentId) {
                 if (isset($sortedItems[$parentId])) {
                     $parentItem = $sortedItems[$parentId];
-    
+
                     $moveHierarchicalList[] = ['id' => $parentItem['id_item'], 'name' => $parentItem['label']];
                     $moveHierarchicalList = array_merge($moveHierarchicalList, buildParentChildTree($parentItem['id_item'], $sortedItems, $menuTree, 1, true));
                 }
             }
-    
+
             return json_encode([
-                'output' => $moveHierarchicalList,  
-                'menu_output' => $menuHierarchicalList,  
+                'output' => $moveHierarchicalList,
+                'menu_output' => $menuHierarchicalList,
                 'selected' => ''
             ]);
         }
     }
-    
+
